requirements.txt
flask
flask-cors
moviepy
pillow
yt-dlp
numpy
torch
whisper
python-dotenv
mistralai